TMS-7-2
There are classes such as "Circle", "Triangle", "Rectangle" that implemented interface "Figure" which contains methods to calculate perimeter, square and show information about figure.

There is "FigureOperationImpl" which implemented interface "FigureOperation" which contains method to calculate summation of perimeters in the array. Besides, it contains method to show information about figures which have been placed in the array.